test: Refactor my company e2e tests
