+++ projects/storefrontapp-e2e-cypress/cypress/helpers/b2b/my-company/config/budget.config.ts
@@ -1,6 +1,6 @@
 import { FULL_BASE_URL_EN_USD } from '../../../site-context-selector';
 import { randomString } from '../../../user';
-import { INPUT_TYPE, MyCompanyConfig } from '../models';
+import { INPUT_TYPE, MyCompanyConfig, MY_COMPANY_FEATURE } from '../models';
 
 export const budgetConfig: MyCompanyConfig = {
   name: 'Budget',
@@ -103,5 +103,10 @@ export const budgetConfig: MyCompanyConfig = {
       objectType: 'costCenters',
       apiEndpoint: '**/constcenters**',
     },
+  ],
+  features: [
+    MY_COMPANY_FEATURE.FORM,
+    MY_COMPANY_FEATURE.LIST,
+    MY_COMPANY_FEATURE.ASSIGNMENTS,
   ],
 };
