+++ projects/storefrontapp-e2e-cypress/cypress/helpers/b2b/my-company/config/purchase-limit.ts
@@ -1,6 +1,6 @@
 import { FULL_BASE_URL_EN_USD } from '../../../site-context-selector';
 import { randomString } from '../../../user';
-import { INPUT_TYPE, MyCompanyConfig } from '../models';
+import { INPUT_TYPE, MyCompanyConfig, MY_COMPANY_FEATURE } from '../models';
 
 export const purchaseLimitConfigs: MyCompanyConfig[] = [
   {
@@ -57,6 +57,7 @@ export const purchaseLimitConfigs: MyCompanyConfig[] = [
         showInDetails: true,
       },
     ],
+    features: [MY_COMPANY_FEATURE.FORM, MY_COMPANY_FEATURE.LIST, ,],
   },
   {
     name: 'Purchase Limit',
@@ -65,7 +66,6 @@ export const purchaseLimitConfigs: MyCompanyConfig[] = [
     apiEndpoint: '/users/current/orderApprovalPermissions',
     objectType: 'orderApprovalPermissions',
     selectOptionsEndpoint: '*orderApprovalPermissionTypes*',
-    disableListChecking: true,
     rows: [
       {
         label: 'Code',
@@ -132,6 +132,7 @@ export const purchaseLimitConfigs: MyCompanyConfig[] = [
         showInDetails: true,
       },
     ],
+    features: [MY_COMPANY_FEATURE.FORM],
   },
   {
     name: 'Purchase Limit',
@@ -140,7 +141,6 @@ export const purchaseLimitConfigs: MyCompanyConfig[] = [
     apiEndpoint: '/users/current/orderApprovalPermissions',
     objectType: 'orderApprovalPermissions',
     selectOptionsEndpoint: '*orderApprovalPermissionTypes*',
-    disableListChecking: true,
     rows: [
       {
         label: 'Code',
@@ -217,5 +217,6 @@ export const purchaseLimitConfigs: MyCompanyConfig[] = [
         showInDetails: true,
       },
     ],
+    features: [MY_COMPANY_FEATURE.FORM],
   },
 ];
