+++ projects/storefrontapp-e2e-cypress/cypress/helpers/b2b/my-company/features/list.ts
@@ -4,51 +4,43 @@ import {
   MyCompanyConfig,
   MyCompanyRowConfig,
   TestListOptions,
-} from './models/index';
-import { loginAsMyCompanyAdmin, waitForData } from './my-company.utils';
+} from '../models/index';
+import { loginAsMyCompanyAdmin, waitForData } from '../my-company.utils';
 
 let requestData: any;
 
-export function testListFromConfig(config: MyCompanyConfig): void {
-  if (!config.disableListChecking) {
-    describe(`${config.name} List`, () => {
-      beforeEach(() => {
-        loginAsMyCompanyAdmin();
-        cy.server();
+export function listTest(config: MyCompanyConfig): void {
+  describe(`${config.name} List`, () => {
+    beforeEach(() => {
+      loginAsMyCompanyAdmin();
+      cy.server();
+    });
+
+    if (!config.nestedTableRows) {
+      it('should show and paginate list', () => {
+        cy.visit(`/organization`);
+        testList(config, {
+          trigger: () =>
+            cy.get(`cx-page-slot.BodyContent a`).contains(config.name).click(),
+        });
       });
 
-      if (!config.nestedTableRows) {
-        it('should show and paginate list', () => {
-          cy.visit(`/organization`);
-          testList(config, {
-            trigger: () =>
-              cy
-                .get(`cx-page-slot.BodyContent a`)
-                .contains(config.name)
-                .click(),
-          });
+      testListSorting(config);
+    } else {
+      it('should show expanded nested list', () => {
+        cy.visit(`/organization`);
+        testList(config, {
+          trigger: () =>
+            cy.get(`cx-page-slot.BodyContent a`).contains(config.name).click(),
+          nested: { expandAll: true },
         });
+      });
 
-        testListSorting(config);
-      } else {
-        it('should show expanded nested list', () => {
-          cy.visit(`/organization`);
-          testList(config, {
-            trigger: () =>
-              cy
-                .get(`cx-page-slot.BodyContent a`)
-                .contains(config.name)
-                .click(),
-            nested: { expandAll: true },
-          });
-        });
-
-        it('should show collapsed nested list', () => {
-          testList(config, { nested: { collapseAll: true } });
-        });
-      }
-    });
-  }
+      it('should show collapsed nested list', () => {
+        testList(config, { nested: { collapseAll: true } });
+      });
+    }
+  });
 }
 
 export function testList(
