+++ projects/storefrontapp-e2e-cypress/cypress/helpers/b2b/my-company/models/my-company.model.ts
@@ -33,6 +33,9 @@ export interface TestListOptions {
 }
 
 export enum MY_COMPANY_FEATURE {
+  LIST = 'list',
+  FORM = 'form',
+  ASSIGNMENTS = 'assignments',
   USER_PASSWORD = 'userPassword',
 }
