+++ projects/storefrontapp-e2e-cypress/cypress/helpers/b2b/my-company/my-company-features.ts
@@ -1,7 +1,13 @@
+import { assignmentsTest } from './features/assignments';
+import { formTest } from './features/form';
+import { listTest } from './features/list';
 import { userPasswordTest } from './features/user-password';
 import { MyCompanyConfig, MY_COMPANY_FEATURE } from './models';
 
 const testMapping = {
+  [MY_COMPANY_FEATURE.LIST]: listTest,
+  [MY_COMPANY_FEATURE.FORM]: formTest,
+  [MY_COMPANY_FEATURE.ASSIGNMENTS]: assignmentsTest,
   [MY_COMPANY_FEATURE.USER_PASSWORD]: userPasswordTest,
 };
